+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 *Release date: Sep 10, 2018*
 
 - **ADDED:** added the ``SPEC_URL`` setting for controlling the download link in ``swagger-ui`` and ``ReDoc``
+- **ADDED:** updated ``ReDoc`` settings (added ``NATIVE_SCROLLBARS`` and ``REQUIRED_PROPS_FIRST``)
 - **IMPROVED:** updated ``swagger-ui`` to version 3.18.2
 - **IMPROVED:** updated ``ReDoc`` to version 2.0.0-alpha.37
 - **FIXED:** stopped generating invalid OpenAPI by improper placement of ``readOnly`` Schemas